Update2 TicTacToe_First_Project.py
In the final iteration I changed and added a few more things that were missing in order for the game to run perfectly. 
I changed the stalemate condition to check after checking for winners because I was getting a 'STALEMATE' ending when a player was winning the game with the last space available. 
I also made the player function a bit more complex by using variable-variable names in order to have one function that works for both players. 
Finally, I made the game replayable at the end and added docstrings to explain what everything does.